# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Web Panel API",
    "version": "1.0.0",
    "description": "Cookie tabanli oturum (express-session) + CSRF korumali API. /api altindaki mutating (POST/PUT/DELETE) endpoint'ler icin CSRF header gonderilmelidir."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "whatsapp.sid"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN"
      }
    },
    "parameters": {
      "AccountIdHeader": {
        "name": "X-Account-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Multi-account icin aktif hesabi secmekte kullanilir."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "Tags list",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Tags", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Tag create",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "color": { "type": ["string", "null"] } } } } } },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/tags/{tagId}": {
      "put": {
        "tags": ["Tags"],
        "summary": "Tag update",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "color": { "type": ["string", "null"] } } } } } },
        "responses": {
          "200": {
            "description": "Success",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["Tags"],
        "summary": "Tag delete",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } }
          }
        }
      }
    },
    "/api/chats/{chatId}/tags": {
      "get": {
        "tags": ["Chats"],
        "summary": "Chat tags",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tags", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Chats"],
        "summary": "Chat tag ekle",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tag_id"],
                "properties": { "tag_id": { "type": "integer" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/chats/{chatId}/tags/{tagId}": {
      "delete": {
        "tags": ["Chats"],
        "summary": "Chat tag kaldir",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/contacts/{chatId}/tags": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Contact tags",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Tags", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Contact tag ekle",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tag_id"],
                "properties": { "tag_id": { "type": "integer" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/contacts/{chatId}/tags/{tagId}": {
      "delete": {
        "tags": ["Contacts"],
        "summary": "Contact tag kaldir",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "chatId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
    "/api/scheduled": {
      "get": {
        "tags": ["Scheduled"],
        "summary": "Scheduled messages list",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Scheduled", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Scheduled"],
        "summary": "Scheduled message create",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["chat_id", "scheduled_at"],
                "properties": {
                  "chat_id": { "type": "string" },
                  "chat_name": { "type": "string" },
                  "message": { "type": "string" },
                  "template_id": { "type": ["integer", "null"] },
                  "scheduled_at": { "type": "string", "format": "date-time" },
                  "is_recurring": { "type": "boolean" },
                  "cron_expression": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/scheduled/{id}": {
      "delete": {
        "tags": ["Scheduled"],
        "summary": "Scheduled message delete",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/api/scripts": {
      "get": {
        "tags": ["Scripts"],
        "summary": "Scripts list (admin)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Scripts", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Scripts"],
        "summary": "Script create (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "code"], "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "code": { "type": "string" }, "trigger_type": { "type": "string" }, "trigger_filter": { "type": ["object", "null"] }, "is_active": { "type": "boolean" } } } } } },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/scripts/test": {
      "post": {
        "tags": ["Scripts"],
        "summary": "Script test (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["code"], "properties": { "code": { "type": "string" }, "testData": { "type": ["object", "null"] } } } } } },
        "responses": { "200": { "description": "Result", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/scripts/{id}": {
      "get": {
        "tags": ["Scripts"],
        "summary": "Script get (admin)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Script", "content": { "application/json": { "schema": { "type": "object" } } } }, "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      },
      "put": {
        "tags": ["Scripts"],
        "summary": "Script update (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      },
      "delete": {
        "tags": ["Scripts"],
        "summary": "Script delete (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/api/scripts/{id}/toggle": {
      "post": {
        "tags": ["Scripts"],
        "summary": "Script toggle (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Result", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/scripts/{id}/run": {
      "post": {
        "tags": ["Scripts"],
        "summary": "Script run (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "type": "object", "properties": { "testData": { "type": "object" } } } } } },
        "responses": { "200": { "description": "Result", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/scripts/{id}/logs": {
      "get": {
        "tags": ["Scripts"],
        "summary": "Script logs (admin)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Logs", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      }
    },
...
  }
}